+++ app/admin/user_documents.rb
@@ -116,15 +116,14 @@ ActiveAdmin.register UserDocument do
   end
 
   form do |f|
+    f.inputs 'Admin' do
+      f.input :organization
+      f.input :uuid
+      f.input :sign_checksum
+    end if current_admin_user.is_root?
     f.inputs do
       f.semantic_errors *f.object.errors.keys
-      if current_admin_user.is_root?
-        f.input :organization
-        f.input :uuid
-        f.input :sign_checksum
-      else
-        f.input :organization, collection: [current_admin_user.organization]
-      end
+      f.input(:organization, collection: [current_admin_user.organization]) unless current_admin_user.is_root?
       tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
       f.input :user_id, as: :search_select, url: admin_users_path(tmp_params),
               fields: [:firstname, :lastname], display_name: :full_name, minimum_input_length: 3,