+++ app/javascript/users/mixins/current_user_mixin.js
@@ -2,7 +2,7 @@
 export const currentUserMixin = {
   computed: {
     isCurrentUserCompleted() {
-      return !!this.currentUser && !!this.currentUser.phone_prefix && !!this.currentUser.phone
+      return !!this.currentUser && !!this.currentUser.phone && !!this.currentUser.birthdate
     },
   },
 }