+++ app/javascript/users/pages/terms_and_conditions.vue
@@ -1,0 +1,187 @@
+<template>
+  <v-container class="justify-center">
+    <h1 class="subtitle"> Termini e condizioni </h1>
+    <p class="subtitle text-center"> Ultimo aggiornamento: 11 settembre 2020 </p>
+    <p> Leggi attentamente questi termini e condizioni prima di utilizzare il nostro servizio. </p>
+    <h1> Interpretazione e definizioni </h1>
+    <h2> Interpretazione </h2>
+    <p> Le parole le cui iniziali sono in maiuscolo hanno significati definiti nelle seguenti condizioni.
+      Le seguenti definizioni avranno lo stesso significato indipendentemente dal fatto che appaiano al singolare o al plurale. </p>
+    <h2> Definizioni </h2>
+    <p> Ai fini dei presenti Termini e condizioni: </p>
+    <ul>
+      <li>
+        <p> <strong> Affiliato </strong> indica un'entità che controlla, è controllata da o è sotto il controllo comune
+          di una parte, dove "controllo" indica la proprietà del 50% o più delle azioni, partecipazioni o altri titoli
+          aventi diritto votare per l'elezione di amministratori o altra autorità di gestione. </p>
+      </li>
+      <li>
+        <p> <strong> Paese </strong> si riferisce a: Italia </p>
+      </li>
+      <li>
+        <p> <strong> Azienda </strong> (denominata "la Società", "Noi", "Noi" o "Nostro" nel presente Contratto) si riferisce a Insiemento. </p>
+      </li>
+      <li>
+        <p> <strong> Dispositivo </strong> indica qualsiasi dispositivo in grado di accedere al Servizio come un computer, un cellulare o un tablet digitale. </p>
+      </li>
+      <li>
+        <p> <strong> Servizio </strong> si riferisce al sito web. </p>
+      </li>
+      <li>
+        <p> <strong> Termini e condizioni </strong> (indicati anche come "Termini") indica i presenti Termini e condizioni
+          che costituiscono l'intero accordo tra l'Utente e la Società per quanto riguarda l'uso del Servizio.</p>
+      </li>
+      <li>
+        <p> <strong> Servizio di social media di terze parti </strong> indica qualsiasi servizio o contenuto (inclusi dati,
+          informazioni, prodotti o servizi) fornito da una terza parte che può essere visualizzato, incluso o reso disponibile dal Servizio. </p>
+      </li>
+      <li>
+        <p> <strong> Sito web </strong> fa riferimento a Insiemento, accessibile da
+          <a href="https://www.insiemento.com" rel="external nofollow noopener" target="_blank">https://www.insiemento.com</a></p>
+      </li>
+      <li>
+        <p> <strong> Tu </strong> indica la persona fisica che accede o utilizza il Servizio, o la società, o altra
+          entità giuridica per conto della quale tale individuo accede o utilizza il Servizio, a seconda dei casi. </p>
+      </li>
+    </ul>
+
+    <h1> Presa visione </h1>
+    <p> Questi sono i termini e le condizioni che regolano l'uso di questo servizio e l'accordo che opera tra te e la società.
+      I presenti Termini e condizioni definiscono i diritti e gli obblighi di tutti gli utenti riguardo all'uso del Servizio. </p>
+    <p> L'accesso e l'utilizzo del Servizio sono subordinati all'accettazione e al rispetto dei presenti Termini e condizioni.
+      I presenti Termini e condizioni si applicano a tutti i visitatori, utenti e altri che accedono o utilizzano il Servizio. </p>
+    <p> Accedendo o utilizzando il Servizio accetti di essere vincolato dai presenti Termini e condizioni. Se non sei
+      d'accordo con qualsiasi parte di questi Termini e condizioni, non puoi accedere al Servizio. </p>
+    <p> Dichiari di avere più di 14 anni. La Società non consente ai minori di 14 anni di utilizzare il Servizio. </p>
+    <p> Il tuo accesso e utilizzo del Servizio è anche condizionato all'accettazione e al rispetto della Privacy Policy
+      della Società. La nostra Informativa sulla privacy descrive le nostre politiche e procedure sulla raccolta, l'uso
+      e la divulgazione delle tue informazioni personali quando usi l'Applicazione o il sito web e ti informa sui
+      tuoi diritti alla privacy e su come la legge ti protegge. Si prega di leggere attentamente la nostra politica
+      sulla privacy prima di utilizzare il nostro servizio. </p>
+
+    <h1> Registrazione </h1>
+    <p>Per poter utilizzare le funzionalità del servizio, devi registrarti fornendo, in maniera
+      veritiera e completa, i dati richiesti in fase di registrazione ed accettare integralmente la
+      <v-btn text class="" :to='{name: "Privacy"}'>Privacy Policy</v-btn>
+      e le presenti condizioni. Hai l’onere di custodire le tue credenziali di accesso.</p>
+    <p>Resta inteso che in nessun caso la Società potrà essere ritenuta responsabile in caso di
+      smarrimento, diffusione, furto o utilizzo non consentito da parte di terzi, a qualsivoglia titolo,
+      delle tue credenziali di accesso.</p>
+
+    <h1> Cancellazione account </h1>
+    <p>Puoi interrompere l’utilizzo del servizio in qualsiasi momento, cancellando il tuo account dall'applicazione o
+      richiederlo alla Società tramite le informazioni di contatto in fondo a questa pagina.</p>
+
+    <h1> Link ad altri siti web </h1>
+    <p> Il nostro Servizio può contenere collegamenti a siti Web o servizi di terze parti che non sono di proprietà
+      o controllati dalla Società. </p>
+    <p> La Società non ha alcun controllo e non si assume alcuna responsabilità per il contenuto, le politiche sulla
+      privacy o le pratiche di siti Web o servizi di terzi. Riconosci inoltre e accetti che la Società non sarà
+      responsabile o responsabile, direttamente o indirettamente, per eventuali danni o perdite causati o presumibilmente
+      causati da o in connessione con l'uso o l'affidamento su tali contenuti, beni o servizi disponibili su o tramite
+      tali siti o servizi web. </p>
+    <p> Ti consigliamo vivamente di leggere i termini e le condizioni e le politiche sulla privacy di qualsiasi sito
+      web o servizio di terze parti che visiti. </p>
+
+    <h1> Risoluzione </h1>
+    <p> Potremmo interrompere o sospendere il tuo accesso immediatamente, senza preavviso o responsabilità, per
+      qualsiasi motivo, incluso, senza limitazione, in caso di violazione dei presenti Termini e condizioni. </p>
+    <p> In caso di risoluzione, il diritto dell'Utente di utilizzare il Servizio cesserà immediatamente. </p>
+    <h1> Limitazione di responsabilità </h1>
+    <p> Nonostante eventuali danni in cui potresti incorrere, l'intera responsabilità della Società e dei suoi fornitori
+      ai sensi di qualsiasi disposizione dei presenti Termini e il tuo esclusivo rimedio per tutto quanto sopra sarà
+      limitata all'importo effettivamente pagato da te tramite il Servizio o 10 € se non hai acquistato nulla tramite il Servizio. </p>
+    <p> Nella misura massima consentita dalla legge applicabile, in nessun caso la Società oi suoi fornitori saranno
+      responsabili per danni speciali, incidentali, indiretti o consequenziali di qualsiasi tipo (inclusi, ma non
+      limitati a, danni per perdita di profitti, perdita di dati o altre informazioni, per interruzione dell'attività,
+      lesioni personali, perdita della privacy derivante da o in qualsiasi modo correlata all'uso o all'impossibilità
+      di utilizzare il Servizio, software di terze parti e / o hardware di terze parti utilizzato con il Servizio,
+      o in altro modo in relazione a qualsiasi disposizione dei presenti Termini), anche se la Società o qualsiasi
+      fornitore è stato informato della possibilità di tali danni e anche se il rimedio fallisce nel suo scopo essenziale. </p>
+    <p> Alcuni stati non consentono l'esclusione di garanzie implicite o la limitazione di responsabilità per danni
+      incidentali o consequenziali, il che significa che alcune delle limitazioni di cui sopra potrebbero non essere
+      applicabili. In questi stati, la responsabilità di ciascuna parte sarà limitata nella misura massima consentita dalla legge. </p>
+
+    <h1> Disclaimer "COSÌ COM'È" e "COME DISPONIBILE" </h1>
+    <p> Il Servizio viene fornito "COSÌ COM'È" e "COME DISPONIBILE" e con tutti i vizi e difetti senza garanzie di alcun tipo.
+      Nella misura massima consentita dalla legge applicabile, la Società, per proprio conto e per conto delle sue
+      Affiliate e dei suoi e dei rispettivi licenziatari e fornitori di servizi, declina espressamente tutte le garanzie,
+      espresse, implicite, legali o di altro tipo, in relazione al Servizio, comprese tutte le garanzie implicite di
+      commerciabilità, idoneità per uno scopo particolare, titolo e non violazione, e garanzie che possono derivare da
+      operazioni commerciali, prestazioni, utilizzo o pratiche commerciali. Senza limitazione a quanto sopra, la Società
+      non fornisce alcuna garanzia o impegno e non garantisce di alcun tipo che il Servizio soddisferà i requisiti
+      dell'utente, raggiungerà i risultati previsti, sarà compatibile o funzionerà con qualsiasi altro software,
+      applicazione, sistema o servizio, senza interruzioni, soddisfano gli standard di prestazioni o affidabilità o sono
+      privi di errori o che eventuali errori o difetti possono o saranno corretti. </p>
+    <p> Fermo restando quanto precede, né la Società né alcuno dei fornitori della società si assume alcuna responsabilità</p>
+
+    <h1> Separabilità e rinuncia </h1>
+    <h2> Separabilità </h2>
+    <p> Se una qualsiasi disposizione dei presenti Termini è ritenuta inapplicabile o non valida, tale disposizione
+      sarà modificata e interpretata per raggiungere gli obiettivi di tale disposizione nella misura più ampia possibile
+      ai sensi della legge applicabile e le restanti disposizioni continueranno ad avere pieno vigore ed effetto. </p>
+    <h2> Rinuncia </h2>
+    <p> Ad eccezione di quanto previsto nel presente documento, il mancato esercizio di un diritto o il mancato rispetto
+      di un obbligo ai sensi dei presenti Termini non influirà sulla capacità di una parte di esercitare tale diritto o
+      richiedere tale prestazione in qualsiasi momento successivo né sarà la rinuncia a una violazione costituiscono una
+      rinuncia a qualsiasi violazione successiva. </p>
+    <h1> Interpretazione della traduzione </h1>
+    <p> I presenti Termini e condizioni potrebbero essere stati tradotti se li abbiamo resi disponibili sul nostro Servizio.
+      Accetti che il testo originale inglese prevarrà in caso di controversia. </p>
+
+    <h1> Modifiche ai presenti Termini e condizioni </h1>
+    <p> Ci riserviamo il diritto, a nostra esclusiva discrezione, di modificare o sostituire i presenti Termini in
+      qualsiasi momento. Se una revisione è materiale Faremo ogni ragionevole sforzo per fornire un preavviso di almeno
+      30 giorni prima che i nuovi termini entrino in vigore. Ciò che costituisce una modifica sostanziale sarà determinato
+      a nostra esclusiva discrezione. </p>
+    <p> Continuando ad accedere o utilizzare il nostro servizio dopo che tali revisioni diventano effettive, accetti di
+      essere vincolato dai termini modificati. Se non accetti i nuovi termini, in tutto o in parte, interrompi l'utilizzo
+      del sito Web e del Servizio. </p>
+
+    <h1> Contattaci </h1>
+    <p> Se hai domande sui presenti Termini e condizioni, puoi contattarci: </p>
+    <ul>
+      <li>
+        <p> Tramite e-mail: privacy@insiemento.com </p>
+      </li>
+    </ul>
+
+    <div class="text-center">
+      <v-btn :to='{name: "home"}' color="primary">Si, accetto</v-btn>
+      <v-btn href="https://www.google.it">Non accetto</v-btn>
+    </div>
+    <div class="mt-10"></div>
+    <br>
+    <div class="mt-10"></div>
+  </v-container>
+</template>
+
+<script>
+  export default {
+    components: {
+    },
+    data() {
+      return {
+      }
+    },
+    computed: {
+    },
+    methods: {
+    },
+  }
+</script>
+
+<style lang="scss" scoped>
+  h1, h2, h3 {
+    text-align: center;
+    color: #666666;
+    margin-top: 2em;
+  }
+  p {
+    color: #444444;
+    &.subtitle {
+      text-align: center;
+      color: #888888;
+    }
+  }
+</style>